+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Fixed
 
 - ggplot2 namespace for function calls in `plotClusterStability`
+- Renamed `stable.tree.clusters`, `get.cluster.graph` and `scan.k.modularity`
+- Removed exports of largeVis internals and `get.cluster.graph`
 
 ### Deprecated